+++ power_grid_model_c/power_grid_model/include/power_grid_model/main_core/topology.hpp
@@ -4,8 +4,8 @@
 
 #pragma once
 
+#include "container_queries.hpp"
 #include "state.hpp"
-#include "state_queries.hpp"
 
 #include "../all_components.hpp"
